move instrumented_nsq to main repo
we will need to have our own custom version in the repo,
so that we can change how it does metrics
